Results from Machine Learning Algorithm
As the results from the Random Forest Classifier show, we can see that Height and Weight have a large impact on the chances of a heart attack. However it's important to note that the size of the dataset is very small, and probably does not show the same results as what the whole population does.